#80 Reworking api constant exports
