+++ ApiDefinitions.swift
@@ -1,14 +1,14 @@
 // Collection of constants for all api uris
 
-// This file was generated by 'api_constants_exporter.py' at 'https://github.com/johannesgrothe/Smarthome_System'
-// Do not modify this file, modify 'api_docs/api_specs.json' and export.
+// This file was generated by 'constants_exporter_api_swift.py' at 'https://github.com/johannesgrothe/Smarthome_System'
+// Do not modify this file, modify 'api_docs/api_specs.json' and 'gadget_docs/gadget_specs.json' and export.
 // Every change made will be overwritten at next export.
 
 
-let api_version = SoftwareVersion(1, 1, 3)
+let api_version = SoftwareVersion(major=1, minor=1, bugfix=3)
 
 
-enum ApiUris : String {
+enum ApiURIs : String {
     case info_bridge = "info/bridge"   // Read Bridge Info
     case bridge_update_check = "bridge/update/check"   // Bridge update check
     case bridge_update_execute = "bridge/update/execute"   // Bridge update execute
