+++ api_definitions.h
@@ -1,8 +1,8 @@
 #pragma once
 // Collection of constants for all api uris
 
-// This file was generated by 'api_constants_exporter.py' at 'https://github.com/johannesgrothe/Smarthome_System'
-// Do not modify this file, modify 'api_docs/api_specs.json' and export.
+// This file was generated by 'constants_exporter_api_cpp.py' at 'https://github.com/johannesgrothe/Smarthome_System'
+// Do not modify this file, modify 'api_docs/api_specs.json' and 'gadget_docs/gadget_specs.json' and export.
 // Every change made will be overwritten at next export.
 
 #include <string>
