+++ exporters/def_docstrings.py
@@ -1,0 +1,1 @@
+DOCSTR_API_CONSTANTS = "Collection of constants for all api uris"
