+++ exporters/def_params.py
@@ -1,0 +1,6 @@
+
+# Base path of github repo for file links
+GITHUB_BASE_FILE_URI = "https://github.com/johannesgrothe/Smarthome_System/blob/master/"
+
+# Path to temporary generated files
+PATH_TEMP_DIR = "temp"
