+++ exporters/wiki_exporter_api_endpoints_bridge.py
@@ -1,9 +1,5 @@
-import logging
 import os
-
-from system.exporters.definitions_loader import DefinitionsLoader
-from system.exporters.script_params import GITHUB_BASE_FILE_URI, PATH_FILE_API_CONSTANTS
-from system.exporters.wiki_export import WikiExporter, WikiExporterApiEndpoints
+from system.exporters.wiki_export import WikiExporterApiEndpoints
 from system.utils.markdown_file import *
 
 _api_specs_file = os.path.join("api_docs", "api_specs.json")
@@ -15,7 +11,7 @@ class WikiExporterApiEndpointsBridge(WikiExporterApiEndpoints):
         file = MarkdownFile()
         file.add(MarkdownHeader(self._api_endpoint_def["title"], 1))
         file.add(MarkdownText(self._api_endpoint_def["description"]))
-        file.add(MarkdownText("Latest Api Version: " + self._api_version))
+        file.add(MarkdownText("Latest Api Version: " + str(self._api_version)))
 
         for category, cat_data in self._api_endpoint_def["items"]["bridge"].items():
             cat_name = cat_data["name"]
